requirements.txt
# Core dependencies
nats-py==2.7.2
aiohttp==3.9.5
PyYAML==6.0.1
Flask==3.0.3
python-dotenv==1.0.1
Jinja2==3.1.4
click==8.1.7

# FastAPI for Control Agent
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
dotenv
# Optional dependencies (comment out if not needed)
# asyncio-mqtt==0.16.2
# uvloop==0.19.0